requirements.txt
numpy==1.6.2
PyGithub==1.9.1
peewee==2.0.4
pep8==1.3.3
scikit-learn==0.12.1
scipy==0.11.0
